+++ test/src/main/kotlin/com/appconstructor/test/TestTable.kt
@@ -1,4 +1,4 @@
-package com.appconstructor
+package com.appconstructor.test
 
 import kotlinx.uuid.SecureRandom
 import kotlinx.uuid.UUID
